+++ CHANGELOG.md
@@ -10,7 +10,8 @@ All notable changes will be documented in this file.
 
 ### Changed
 
-* Adding `MarginMode` to `CreateOrder`, `OrderUpdate`, `TradeUpdate`
+* Adding `PositionEffect` to `OrderAck`
+* Adding `MarginMode` to `CreateOrder`, `OrderAck`, `OrderUpdate`, `TradeUpdate`
 * OMS validate quantity and price against reference data (#432)
 * Added `MarginMode` to `FundsUpdate` and `PositionUpdate` (#430)